+++ packages/analytics-js-common/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [3.26.0](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-common@3.25.0...@rudderstack/analytics-js-common@3.26.0) (2025-11-19)
+
+
+### Features
+
+* check storage type availability for events persistence ([#2595](https://github.com/rudderlabs/rudder-sdk-js/issues/2595)) ([dd5bec5](https://github.com/rudderlabs/rudder-sdk-js/commit/dd5bec51e3a5bd8d678d2835259a1cfa1ce2614e))
+
 ## [3.25.0](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-common@3.24.0...@rudderstack/analytics-js-common@3.25.0) (2025-09-15)